More Setup
Really spent a lot of time contemplating the design of the application and prompting chat gpt about Next.js to formulate a plan of action. Really help me focus on the goal and simplified my outlook of what this app is going to be. Just a one page app with a ton of components that will use a search form should be basic just to get me feet wet.